# perf(sql): optimize count_distinct symbol function