--- a/core/src/main/java/io/questdb/griffin/engine/functions/GroupByFunction.java
+++ b/core/src/main/java/io/questdb/griffin/engine/functions/GroupByFunction.java
@@ -40,18 +40,32 @@ public interface GroupByFunction extends Function, Mutable {
 
     void computeNext(MapValue mapValue, Record record);
 
-    default void interpolateBoundary(MapValue mapValue1,
-                                     MapValue mapValue2,
-                                     long boundaryTimestamp,
-                                     boolean isEndOfBoundary) {
+    // only makes sense for non-keyed group by
+    default boolean earlyExit(MapValue mapValue) {
+        return false;
+    }
+
+    default void interpolateBoundary(
+            MapValue mapValue1,
+            MapValue mapValue2,
+            long boundaryTimestamp,
+            boolean isEndOfBoundary
+    ) {
         throw new UnsupportedOperationException();
     }
 
-    default void interpolateGap(MapValue mapValue,
-                                MapValue mapValue1,
-                                MapValue mapValue2,
-                                long x) {
+    default void interpolateGap(
+            MapValue mapValue,
+            MapValue mapValue1,
+            MapValue mapValue2,
+            long x
+    ) {
         throw new UnsupportedOperationException();
+    }
+
+    // only makes sense for non-keyed group by
+    default boolean isEarlyExitSupported() {
+        return false;
     }
 
     default boolean isInterpolationSupported() {