… core/src/test/java/io/questdb/test/ServerMainTest.java
… core/src/test/java/io/questdb/test/griffin/engine/functions/groupby/CountDistinctSymbolGroupByFunctionFactoryTest.java