Includes the following optimizations:
* Replaces `CompactIntHashSet` with `BitSet` in `CountDistinctSymbolGroupByFunction`. The former has a much more compact memory layout and faster operations.
* Introduces early exit for non-keyed queries with `count_distinct()` over symbol columns, e.g. `SELECT count_distinct(sym_col) FROM my_table;`. The exit condition is based on the symbol table sizes: as soon as they're reached, row iteration stops.

```sql
CREATE TABLE x AS (
  SELECT rnd_symbol(1000, 10, 10, 0) sym1,
         rnd_symbol('a', 'b', 'c') sym2,
         timestamp_sequence('2023-11-17T10', 1000000L) ts
  FROM long_sequence(100000000)
) TIMESTAMP(ts) PARTITION BY DAY;

-- before 1s
-- after 10ms
SELECT count_distinct(sym1) FROM x;

-- before 2s
-- after 10ms
SELECT count_distinct(sym1), count_distinct(sym2) FROM x;

-- before 2.5s
-- after 2.4s
SELECT sym2, count_distinct(sym1) FROM x;
```