@ideoma 2023-11-17
#### [PR Coverage check]

:heart_eyes: **pass** : 87 / 88 (98.86%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/griffin/engine/functions/GroupByFunction.java|1|2|50.00%|
|:large_blue_circle:|io/questdb/std/BitSet.java|24|24|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/CountDistinctSymbolGroupByFunction.java|18|18|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/GroupByNotKeyedRecordCursorFactory.java|44|44|100.00%|