@bluestreak01 2023-11-17
@puzpuzpuz this is a very clever improvement 👍 how will it work with `where` clause?