@puzpuzpuz 2023-11-17
> @puzpuzpuz this is a very clever improvement 👍 how will it work with `where` clause?

If the filtered symbol values are only a subset of the full symbol table, the early exit check will be no-op (full loop over the base cursor). If all symbol values are returned by the filter, the check will kick in and redundant iteration over the base cursor will stop. That's done only in the `GroupByNotKeyedRecordCursorFactory`, so all other factories behave in the same way.