@ideoma 2023-11-17
#### [PR Coverage check]

:heart_eyes: **pass** : 92 / 93 (98.92%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/griffin/engine/functions/GroupByFunction.java|1|2|50.00%|
|:large_blue_circle:|io/questdb/std/BitSet.java|24|24|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/CountDistinctSymbolGroupByFunction.java|18|18|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/GroupByNotKeyedRecordCursorFactory.java|49|49|100.00%|