@ideoma 2023-11-20
#### [PR Coverage check]

:heart_eyes: **pass** : 68 / 71 (95.77%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/CountDistinctUuidGroupByFunction.java|0|1|00.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/CountDistinctIPv4GroupByFunction.java|0|1|00.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/GroupByFunction.java|1|2|50.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/CountDistinctStringGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/std/BitSet.java|22|22|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/CountDistinctSymbolGroupByFunction.java|19|19|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/CountDistinctLongGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/CountDistinctLong256GroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/GroupByNotKeyedRecordCursorFactory.java|22|22|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/CountDistinctIntGroupByFunction.java|1|1|100.00%|